Feat: 정책 필터 컴포넌트로 구분
